requirements.txt
django~=4.0.6
djangorestframework~=3.13.1
biopython
Blastpy3
django-cors-headers
